pyproject.toml
[project]
name = "kimi-docs-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "google-genai>=1.62.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.11",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "torch>=2.10.0",
]
